Refactor: "코드리뷰 반영"
매직넘버 제거, data 파일 분리, 클래스 이름 / 메서드 이름 변경
