+++ sass/layout/_side-bar.scss
@@ -1,21 +1,21 @@
 .side-bar {
   position: fixed;
   top: 0;
-  left: -100%;
-  // transform: translateX(-100%);
+  // left: -100%;
+  transform: translateX(-100%);
   height: 100%;
   z-index: 100;
-  transition: left 250ms ease-in-out;
-  // transition: transform 250ms ease-in-out;
+  // transition: left 250ms ease-in-out;
+  transition: transform 250ms ease-in-out;
 
   &.show-side-bar {
-    left: 0;
+    // left: 0;
+    transform: translate(0%);
   }
 
   &__close {
     position: absolute;
     left: 32rem;
-    // transform: translate(0%);
     margin-top: 12px;
     margin-left: 12px;
   }
