+++ scripts/constants.js
@@ -1,0 +1,8 @@
+// POPUP
+export const POPUP_DELAY_MILLI = 1000;
+
+// SLIDER
+export const LEFT = -1;
+export const RIGHT = 1;
+export const AUTO_SLIDE_MILLI = 10000;
+export const SLIDE_TRANSITION_MILLI = 250;
