+++ style.css
@@ -265,12 +265,12 @@ body {
 .side-bar {
   position: fixed;
   top: 0;
-  left: -100%;
+  transform: translateX(-100%);
   height: 100%;
   z-index: 100;
-  transition: left 250ms ease-in-out; }
+  transition: transform 250ms ease-in-out; }
   .side-bar.show-side-bar {
-    left: 0; }
+    transform: translate(0%); }
   .side-bar__close {
     position: absolute;
     left: 32rem;
